Completed 80% of the project.
It can now understand templates, content-files, exclude content and metadata from them, turn that into HTML, and save that HTML. Tested it; it works.

How do they say? That 80% of a given thing is 20% its substance, and 20% of it is the remaining 80%? Ye, sooo... The remaining 20% is CSS generation. Which is the primary reason why this project exists. Oops.